Lottos : 로또 구매 기능 이동
