+++ src/main/java/lotto/LottoControllerMain.java
@@ -1,17 +1,15 @@
 package lotto;
 
 import lotto.domain.*;
-import lotto.util.LottoGenerator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class LottoControllerMain {
     public static void main(String[] args) {
         Money money = new Money(InputView.inputMoney());
-        Lottos lottos = new Lottos(buyLotto(money.countBuyLotto()));
+        Lottos lottos = new Lottos(money.countBuyLotto());
 
         OutputView.showBuyLotto(lottos.getLottos());
 
@@ -23,13 +21,5 @@ public class LottoControllerMain {
         OutputView.resultMessage();
         OutputView.result(Result.getResultValues(), new Statistics(results));
         OutputView.showTotalProfit(money.calculateProfitRate(Result.calculateProfit(results)));
-    }
-
-    public static List<Lotto> buyLotto(int count){
-        List<Lotto> lottos = new ArrayList<>();
-        for (int i = 0; i < count; i++) {
-            lottos.add(new Lotto(LottoGenerator.make()));
-        }
-        return lottos;
     }
 }
